+++ stroom-util/src/main/java/stroom/util/pipeline/scope/PipelineScopeRunnable.java
@@ -57,4 +57,8 @@ public class PipelineScopeRunnable {
 
         return result;
     }
+
+    public boolean isScopeActive() {
+        return scope.isScopeActive();
+    }
 }